Moved states files into the states folder + player
player entity has it's own folder. All future entity types will have
their own folder
