+++ src/entities/index.js
@@ -1,7 +1,7 @@
 import { createAction } from '@reduxjs/toolkit';
 import produce from 'immer';
 
-import { typeDefinitions } from './typeDefinitions';
+import player from './player';
 
 let _id= 0;
 export const createEntity = createAction('CREATE_ENTITY', ({ id, type, state, props }) => ({
@@ -13,9 +13,14 @@ export const createEntity = createAction('CREATE_ENTITY', ({ id, type, state, pr
   },
 }));
 
+export const entityDefinitions = {
+  player,
+};
+
 // put the action(s) that are being called for the update,
 // also the values they're contributing, into this meta object.
 // For debugging purposes.
+// ~~ generic action for all entities
 export const updateProps = createAction('UPDATE_PROPS', (payload, meta) => ({
   payload,
   meta,
@@ -34,7 +39,7 @@ const entitiesReducer = (state = {}, action) => produce(state, draftState => {
       draftState[id] = {
         id,
         type,
-        states: produce(states, states => typeDefinitions[type].stateReducer(states, action)),
+        states: produce(states, states => entityDefinitions[type].stateReducer(states, action)),
         props: { x, y, vx, vy, height, width },
       };
       break;
@@ -50,7 +55,7 @@ const entitiesReducer = (state = {}, action) => produce(state, draftState => {
     default:
       if (action.payload && action.payload.id !== undefined) {
         const entity = draftState[action.payload.id];
-        typeDefinitions[entity.type].stateReducer(entity.states, action);
+        entityDefinitions[entity.type].stateReducer(entity.states, action);
       }
   }
 });
