+++ src/entities/states/createState.js
@@ -1,0 +1,1 @@
+export const createState = (state, extraProps = {}) => ({ [state]: { createdAt: Date.now(), ...extraProps } });
