+++ src/entities/states/jump.js
@@ -1,4 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
+
 import { createState } from './createState';
 
 export const States = {
