+++ src/getLogger/index.js
@@ -24,7 +24,7 @@ const logger = () => {
         const filteredLogs = logs.filter(getFilter(filters));
         // we check lazily if the filtered logs has changed.
         if (oldCount !== filteredLogs.length) {
-          printFirstThreeDiffs(filteredLogs);
+          printType(filteredLogs);
         }
         oldCount = filteredLogs.length;
         return result;
