+++ thingstoconsider.rtf
@@ -52,8 +52,9 @@
  - finish looking at TODOs,adding bullets below\
 \
  *- DRYing diff code\
- - lastState for for graphics stopping isn\'92t being used\
+ *- lastState for for graphics stopping isn\'92t being used\
  - ?entities/index should represent the whole entity, so put reducer in it\'92s own file, and maybe import export stuff???\
+\
  - add button => filter => actions.\
 	- Add actionFilter to typeDefinitions [action filters is a thunk that takes an action]\
 		- typeDefn[player].actionFilter = action => (dispatch, getState) => dispatch(action)\
